+++ queuebie/__init__.py
@@ -1,6 +1,6 @@
 """A simple and synchronous message queue for commands and events for Django"""
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 
 from queuebie.registry import MessageRegistry