+++ testapp/nested_domain/messages/commands/nested_commands.py
@@ -1,0 +1,8 @@
+import dataclasses
+
+from queuebie.messages import Command
+
+
+@dataclasses.dataclass(kw_only=True)
+class DoSomethingNested(Command):
+    my_var: int